Fixed IP management for IPs > 127.255.255.255

+++ GridMapper/GridMapper/Execution/Execution.cs
@@ -44,7 +44,7 @@ namespace GridMapper
 					Parallel.ForEach<int>( _option.IpToTest.Result, new ParallelOptions { MaxDegreeOfParallelism = 200 }, ipInt =>
 						{
 							//PingSender pingSender = new PingSender( Option );
-							IPAddress ip = IPAddress.Parse( ipInt.ToString() );
+							IPAddress ip = IPAddress.Parse( ((uint)ipInt).ToString() );
 							PingReply pingReply = pingSender.Ping( ip );
 							if( pingReply != null ) 
 								_repository.AddOrUpdate( ip, pingReply );
